backend/requirements.txt
httpx
beautifulsoup4
playwright
pyyaml
pydantic
supabase
loguru
python-dotenv
